pyproject.toml
[tool.poetry]
name = "zeabur-extend-trial"
version = "0.1.0"
description = ""
authors = ["bestk <1995kaikai@gmail.com>"]
readme = "README.md"
packages = [{ include = "zeabur_extend_trial" }]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.26.0"
pytest = "^7.4.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
